requirements.txt
aiohttp==3.8.5
aiohttp-retry==2.8.3
aiosignal==1.3.1
asgiref==3.7.2
async-timeout==4.0.2
attrs==23.1.0
backports.zoneinfo==0.2.1
beautifulsoup4==4.12.2
certifi==2023.7.22
charset-normalizer==3.2.0
config==0.5.1
decouple==0.0.7
defusedxml==0.7.1
diff-match-patch==20230430
Django==4.2.2
django-import-export==3.2.0
django-livereload-server==0.4
et-xmlfile==1.1.0
frozenlist==1.4.0
idna==3.4
import-export==0.3.1
livereload==2.6.3
MarkupPy==1.14
multidict==6.0.4
odfpy==1.4.1
openpyxl==3.1.2
Pillow==10.0.0
PyJWT==2.8.0
pyngrok==6.0.0
pypng==0.20220715.0
pytz==2023.3
PyYAML==6.0.1
qrcode==7.4.2
requests==2.31.0
six==1.16.0
soupsieve==2.4.1
sqlparse==0.4.4
tablib==3.5.0
tornado==6.3.2
twilio==8.5.0
typing_extensions==4.7.1
tzdata==2023.3
urllib3==2.0.4
xlrd==2.0.1
xlwt==1.3.0
yarl==1.9.2
